requirements.txt
asgiref==3.11.0
dj-database-url==3.1.2
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
et_xmlfile==2.0.0
gunicorn==25.1.0
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
